+++ linkerd/core/src/main/scala/io/buoyant/linkerd/Linker.scala
@@ -31,10 +31,11 @@ object Linker {
     tlsClient: Seq[TlsClientInitializer] = Nil,
     tracer: Seq[TracerInitializer] = Nil,
     identifier: Seq[IdentifierInitializer] = Nil,
-    classifier: Seq[ResponseClassifierInitializer] = Nil
+    classifier: Seq[ResponseClassifierInitializer] = Nil,
+    announcer: Seq[AnnouncerInitializer] = Nil
   ) {
     def iter: Iterable[Seq[ConfigInitializer]] =
-      Seq(protocol, namer, interpreter, tlsClient, tracer, identifier, classifier)
+      Seq(protocol, namer, interpreter, tlsClient, tracer, identifier, classifier, announcer)
 
     def all: Seq[ConfigInitializer] = iter.flatten.toSeq
 
@@ -52,7 +53,8 @@ object Linker {
     LoadService[TlsClientInitializer],
     LoadService[TracerInitializer],
     LoadService[IdentifierInitializer],
-    LoadService[ResponseClassifierInitializer]
+    LoadService[ResponseClassifierInitializer],
+    LoadService[AnnouncerInitializer]
   )
 
   def parse(